move @view and @api definition to separate modules. Make @api.xxx url parameter positional-only
